Added a Menu UI
I've created a menu UI which allows the user to read instructions, allows them to change settings which can be the volume, activating boss mode, or starting at a different round number. They can also exit the game and start the game as well.

I've added some pictures and titles to enhance the feel of the game and added some music and button clicking effects.