Added Data Persistence between Scenes
I added a Main Manager script such that it would pass the settings data to the next scene which would be the main game.

This includes the boss active bool, the volume value as well as the starting round value.

I also changed the round input from an input field to a slider so it would be easier to pass the value as a number rather than a string.