Added Basic Player Movement
I ensured there is encapsulation of variables and methods for each class.

I also worked on player input and movement. For this this one I managed to work on WASD player movement. I also worked on the player camera and make it into an FPS style game.

I also added a jumping mechanic if the user presses spacebar as well as a sprinting mechanic is the user presses Left-Shift