Player UI and Pause Screen
In this session I added a player UI which includes a health bar as well as a stamina bar.

Whenever the player runs, their stamina will go down and if they walk/stand still, it regenerates. However if it reaches 0, then they have to wait till it reaches 100 again to run once more.

I also added player animations to give it a little bit of realism.

Finally, I added a pause ability by pressing 'P' and this can allow the user to return to the Main Menu or simply restart. However, I have yet to find a way to save data from the Main to the Menu in case the user wants to keep the previous settings. Although I may just also leave it like this as well...