Added a Game Over Feature
I tweaked some of the prefabs for reusability and I also added a game over feature.

When the player dies, they will not be able to move and all the objects in the world (zombies) will despawn. Additionally, they are presented with a screen saying GAME OVER and they can choose to restart or return to the main menu.