Added the health and ammo stations
I added 3 health and 3 ammo stations throughout the map and coded their functionality.

You can go up to them and you can request for more health or ammo if available by pressing 'C'. Then you have to wait every 2 minutes for it to respawn once more.

Will improve this feature with sound effects and particle FX as well as implement it with perk stations which I plan to do next.