Perk Stations added
In this session, I added 4 Perk Stations:
- Armour endurance: reduces the damage inflicted from a zombie on the player

- Stamina Endurance: reduces the rate at which stamina depletes.

- Damage Boost: Increases the damage inflicted on zombies

- Point Multiplier: Increases the number of points when killed a zombie.

I have completed the design and implemented the prefabs onto the scene. I have also coded the trigger events when the player enters the colliders of the perk stations.

I also worked on the corresponding UI pop ups with each collider as well as the UI of the perks on the top left.

I also added a simple point variable so now the player can track their points and spend it on perks.

I only completed the functionality of two perks: armour and stamina as the other two don't have coded implementations yet.

For the armour endurance: it can never be greater than the zombie damage but very close to it and it scales accordingly with zombie damage.

For the stamina endurance, it can never be greater than 20 as that is the max depletion rate but it can be very close to it.

Both of these functions were calculated and experimented with in desmos.