Added a gun and FX
I added a gun prefab and some animations for shooting and reloading.

I added some nice sound effects for them as well as particle effects for when the gun shoots.

I tried working on the connection between the gun shot and the object that it collides with. I managed to make it work by detecting which object it is, however there are still many problems:

1) the zombies are not getting damaged (some non referenced error)

2) There are still errors with the perk stations and their values actually working...

3) There is a problem with the reload where the sound effect can always be played everytime it hits reload and this needs to be fixed for a one time playing when reloading.