Finally fixed the zombies taking damage
I finally fixed the issue where the zombies were not taking any damage. This is because I was not referencing their characterstats each independently but as a whole...

However, the point system is still not working hence, need to fix that. Additionally, I want to add some nice jumping sound effects and zombie death effects and zombie sounds to emulsify the entire experience even more.