Added an ammo system and fixed point system
I added an ammo system with the weapon and tied it to the ammo stations as well. It works perfectly which is nice.

I also fixed the issue with the point system and the problem was we were basing the points based on the player prefab in the files and not the gameobject in the scene itself. Hence I changed this so that whenever a zombie dies, it refers to the player gameobject in the scene's points.